refactor(trusty-common): split github/jira/linear backends into submodules under 500-SLOC cap (#607)
Split three oversized backend files into focused submodules:

- backends/github.rs (721 SLOC) → github/{mod,types,client,backend}.rs
- backends/jira.rs (758 SLOC) → jira/{mod,types,client,backend}.rs
- backends/linear.rs (686 SLOC) → linear/{mod,types,client,backend}.rs

Each backend uses the split convention: types.rs holds the struct and
parse helpers, client.rs holds HTTP transport methods, backend.rs holds
the Backend trait impl and tests. mod.rs is a thin pub-use facade.

Remove the three stale entries from .line-cap-allowlist.tsv.
