+++ crates/trusty-common/src/tickets/api/backends/linear/backend.rs
@@ -1,222 +1,24 @@
-//! Linear backend (GraphQL).
+//! Linear backend — `Backend` trait implementation.
 //!
-//! Why: Linear's API is GraphQL-only; no REST surface.
-//! What: All queries are `const &str`. Single `graphql` helper handles
-//! auth, payload assembly, and error extraction.
-//! Test: shape tests below; live tests gated by env vars.
+//! Why: Isolates the Backend trait methods from transport and parse helpers
+//! so the file stays under the 500-SLOC cap.
+//! What: Implements the full `Backend` trait for `LinearBackend`, using
+//! the GraphQL helper and parse helpers from the sibling submodules.
+//! Test: `tests::priority_to_int_mapping`, `tests::int_to_priority_mapping`,
+//! `tests::state_mapping`, `tests::requires_api_key`.
 
-use anyhow::{Context, Result, anyhow, bail};
+use anyhow::{Result, bail};
 use async_trait::async_trait;
 use chrono::{DateTime, Utc};
-use parking_lot_compat::Mutex;
-use reqwest::Client;
-use serde_json::{Value, json};
+use serde_json::json;
 
 use crate::tickets::api::backends::{
     Backend, CreateIssueParams, CreateMilestoneParams, ListIssuesParams, SearchIssuesParams,
     UpdateIssueParams,
 };
-use crate::tickets::api::config::LinearConfig;
 use crate::tickets::api::models::*;
 
-mod parking_lot_compat {
-    // Why: avoid pulling parking_lot as a dep; std Mutex is fine for the
-    // tiny critical section (team_id memoisation).
-    pub use std::sync::Mutex;
-}
-
-const GRAPHQL_URL: &str = "https://api.linear.app/graphql";
-const USER_AGENT: &str = "trusty-tickets/0.1";
-
-/// Linear GraphQL backend.
-///
-/// Why: API key + lazily-resolved team_id + HTTP client.
-/// What: `team_id` is resolved on first use from `team_key`, cached in a
-/// `Mutex<Option<String>>`.
-/// Test: `tests::priority_to_int_mapping`.
-pub struct LinearBackend {
-    api_key: String,
-    team_key: Option<String>,
-    team_id: Mutex<Option<String>>,
-    http: Client,
-}
-
-impl LinearBackend {
-    /// Why: Accept either `team_id` or `team_key`; key is resolved later.
-    /// What: Construct + validate API key presence.
-    /// Test: `tests::requires_api_key`.
-    pub fn new(cfg: LinearConfig) -> Result<Self> {
-        let api_key = cfg
-            .api_key
-            .ok_or_else(|| anyhow!("linear: missing api_key (set LINEAR_API_KEY)"))?;
-        let http = Client::builder()
-            .user_agent(USER_AGENT)
-            .build()
-            .context("build linear http client")?;
-        Ok(Self {
-            api_key,
-            team_key: cfg.team_key,
-            team_id: Mutex::new(cfg.team_id),
-            http,
-        })
-    }
-
-    async fn graphql(&self, query: &str, variables: Value) -> Result<Value> {
-        let resp = self
-            .http
-            .post(GRAPHQL_URL)
-            .header("Authorization", &self.api_key)
-            .header("Content-Type", "application/json")
-            .json(&json!({ "query": query, "variables": variables }))
-            .send()
-            .await
-            .context("linear graphql")?;
-        let status = resp.status();
-        let text = resp.text().await.context("read body")?;
-        if !status.is_success() {
-            bail!("linear API failed: {status}: {text}");
-        }
-        let v: Value =
-            serde_json::from_str(&text).with_context(|| format!("parse json: {text}"))?;
-        if let Some(errors) = v.get("errors") {
-            bail!("linear graphql errors: {errors}");
-        }
-        Ok(v)
-    }
-
-    async fn resolve_team_id(&self) -> Result<String> {
-        {
-            let g = self.team_id.lock().unwrap();
-            if let Some(t) = &*g {
-                return Ok(t.clone());
-            }
-        }
-        let key = self
-            .team_key
-            .clone()
-            .ok_or_else(|| anyhow!("linear: missing team_key/team_id"))?;
-        let q = "query($key: String!) { team(id: $key) { id name key } }";
-        // Linear's `team(id:)` actually accepts the team key too; if that
-        // fails we fall back to scanning teams.
-        let mut id_opt = None;
-        if let Ok(v) = self.graphql(q, json!({ "key": key })).await
-            && let Some(id) = v["data"]["team"]["id"].as_str()
-        {
-            id_opt = Some(id.to_string());
-        }
-        if id_opt.is_none() {
-            let q2 = "query { teams(first: 100) { nodes { id key name } } }";
-            let v = self.graphql(q2, json!({})).await?;
-            let nodes = v["data"]["teams"]["nodes"]
-                .as_array()
-                .cloned()
-                .unwrap_or_default();
-            for n in nodes {
-                if n.get("key").and_then(|k| k.as_str()) == Some(&key) {
-                    id_opt = n.get("id").and_then(|v| v.as_str()).map(String::from);
-                    break;
-                }
-            }
-        }
-        let id = id_opt.ok_or_else(|| anyhow!("linear: team '{key}' not found"))?;
-        *self.team_id.lock().unwrap() = Some(id.clone());
-        Ok(id)
-    }
-}
-
-fn priority_to_int(p: &str) -> i32 {
-    match p {
-        "critical" => 1,
-        "high" => 2,
-        "medium" => 3,
-        "low" => 4,
-        _ => 0,
-    }
-}
-
-fn int_to_priority(n: i64) -> Option<Priority> {
-    match n {
-        1 => Some(Priority::Critical),
-        2 => Some(Priority::High),
-        3 => Some(Priority::Medium),
-        4 => Some(Priority::Low),
-        _ => None,
-    }
-}
-
-fn state_from_name(s: &str) -> IssueState {
-    let l = s.to_lowercase();
-    match l.as_str() {
-        "in progress" | "in_progress" => IssueState::InProgress,
-        "ready" => IssueState::Ready,
-        "tested" => IssueState::Tested,
-        "done" | "completed" => IssueState::Done,
-        "blocked" => IssueState::Blocked,
-        "waiting" => IssueState::Waiting,
-        "canceled" | "cancelled" | "closed" => IssueState::Closed,
-        _ => IssueState::Open,
-    }
-}
-
-fn parse_issue(node: &Value) -> Issue {
-    let id = node["id"].as_str().unwrap_or("").to_string();
-    let state_name = node["state"]["name"].as_str().unwrap_or("Backlog");
-    let labels = node["labels"]["nodes"]
-        .as_array()
-        .map(|a| {
-            a.iter()
-                .filter_map(|l| l.get("name").and_then(|n| n.as_str()).map(String::from))
-                .collect()
-        })
-        .unwrap_or_default();
-    let assignee = node["assignee"]["name"].as_str().map(String::from);
-    let priority = node["priority"].as_i64().and_then(int_to_priority);
-    Issue {
-        id,
-        backend: "linear".into(),
-        url: node["url"].as_str().map(String::from),
-        title: node["title"].as_str().unwrap_or("").to_string(),
-        description: node["description"].as_str().map(String::from),
-        state: state_from_name(state_name),
-        issue_type: IssueType::Issue,
-        priority,
-        assignee,
-        labels,
-        milestone_id: node["cycle"]["id"].as_str().map(String::from),
-        milestone_name: node["cycle"]["name"].as_str().map(String::from),
-        project_id: node["project"]["id"].as_str().map(String::from),
-        project_name: node["project"]["name"].as_str().map(String::from),
-        parent_id: node["parent"]["id"].as_str().map(String::from),
-        children: node["children"]["nodes"]
-            .as_array()
-            .map(|a| {
-                a.iter()
-                    .filter_map(|c| c.get("id").and_then(|v| v.as_str()).map(String::from))
-                    .collect()
-            })
-            .unwrap_or_default(),
-        created_at: node["createdAt"]
-            .as_str()
-            .and_then(|s| DateTime::parse_from_rfc3339(s).ok())
-            .map(|d| d.with_timezone(&Utc)),
-        updated_at: node["updatedAt"]
-            .as_str()
-            .and_then(|s| DateTime::parse_from_rfc3339(s).ok())
-            .map(|d| d.with_timezone(&Utc)),
-        extra: node.clone(),
-    }
-}
-
-const ISSUE_FIELDS: &str = r#"
-    id identifier title description priority url createdAt updatedAt
-    state { name }
-    assignee { name }
-    labels { nodes { name } }
-    cycle { id name }
-    project { id name }
-    parent { id }
-    children { nodes { id } }
-"#;
+use super::types::{ISSUE_FIELDS, LinearBackend, parse_issue, priority_to_int};
 
 #[async_trait]
 impl Backend for LinearBackend {
@@ -727,7 +529,10 @@ impl Backend for LinearBackend {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use super::super::types::{int_to_priority, priority_to_int, state_from_name};
+    use crate::tickets::api::backends::linear::LinearBackend;
+    use crate::tickets::api::config::LinearConfig;
+    use crate::tickets::api::models::{IssueState, Priority};
 
     #[test]
     fn priority_to_int_mapping() {
